GG-8781 fixing generator's tests
