+++ build.sbt
@@ -3,16 +3,14 @@ ThisBuild / majorVersion := 13
 ThisBuild / isPublicArtefact := true
 
 lazy val domain = (project in file("."))
-  .settings(
-    scalacOptions ++= Seq(
-      "-Werror"
-    ),
-    publish / skip := true
-  )
+  .settings(publish / skip := true)
   .aggregate(domainPlay30, domainTestPlay30)
 
 lazy val domainPlay30 = Project("domain-play-30", file("domain-play-30"))
   .settings(
+    scalacOptions ++= Seq(
+      "-Werror"
+    ),
     libraryDependencies ++= LibDependencies(),
     scalafmtOnCompile := true
   )
@@ -20,6 +18,9 @@ lazy val domainPlay30 = Project("domain-play-30", file("domain-play-30"))
 
 lazy val domainTestPlay30 = Project("domain-test-play-30", file("domain-test-play-30"))
   .settings(
+    scalacOptions ++= Seq(
+      "-Werror"
+    ),
     libraryDependencies ++= LibDependencies(),
     scalafmtOnCompile := true
   )
